feat: debug animation and keep animation images v1.1.8

+++ src/util/AnimationGeneration.ts
@@ -3,16 +3,19 @@ import { Executor, OpenScadOutputWithSummary } from "openscad-cli-wrapper";
 export async function GenerateWebpAnimation(
   output: OpenScadOutputWithSummary,
   animDelay: number,
+  debug: boolean,
   executor: Executor,
 ): Promise<OpenScadOutputWithSummary> {
   const animImagesPattern = output.file;
-  output.file = animImagesPattern.replace("_animImg*.png", ".webp").replace("_animation", "");
-  const debug = ""; // "-v"
-  output.output += await executor(`img2webp ${debug} -o "${output.file}" -d "${animDelay}" ${animImagesPattern}`);
-  try {
-    output.output += await executor(`rm ${animImagesPattern}`);
-  } catch (error) {
-    console.log(`Error cleaning anim images (${animImagesPattern})`, error);
+  output.file = animImagesPattern.replace("_animImg*.png", ".webp");
+  const debugCmd = debug ? "-v" : "";
+  output.output += await executor(`img2webp ${debugCmd} -o "${output.file}" -d "${animDelay}" ${animImagesPattern}`);
+  if (!debug) {
+    try {
+      output.output += await executor(`rm ${animImagesPattern}`);
+    } catch (error) {
+      console.log(`Error cleaning anim images (${animImagesPattern})`, error);
+    }
   }
   return output;
 }
@@ -20,18 +23,21 @@ export async function GenerateWebpAnimation(
 export async function GenerateGifAnimation(
   output: OpenScadOutputWithSummary,
   animDelay: number,
+  debug: boolean,
   executor: Executor,
 ): Promise<OpenScadOutputWithSummary> {
   const animImagesPattern = output.file;
-  output.file = animImagesPattern.replace("_animImg*.png", ".gif").replace("_animation", "");
-  const debug = ""; // "-verbose"
+  output.file = animImagesPattern.replace("_animImg*.png", ".gif");
+  const debugCmd = debug ? "-verbose" : "";
   output.output += await executor(
-    `convert ${debug} -delay "${animDelay * 10}" -loop 0 ${animImagesPattern} "${output.file}"`,
+    `convert ${debugCmd} -delay "${animDelay * 10}" -loop 0 ${animImagesPattern} "${output.file}"`,
   );
-  try {
-    output.output += await executor(`rm ${animImagesPattern}`);
-  } catch (error) {
-    console.log(`Error cleaning anim images (${animImagesPattern})`, error);
+  if (!debug) {
+    try {
+      output.output += await executor(`rm ${animImagesPattern}`);
+    } catch (error) {
+      console.log(`Error cleaning anim images (${animImagesPattern})`, error);
+    }
   }
   return output;
 }
